feat - Order Features According To Correlation
The order of the dropdown feature selection is based on the correlation
between the features. The next feature in the sequence is the feature
most correlated to the previous feature and that hasn't already been
placed in the sequence. The dimensions of the parallel coordinates plot
will appear in this order whenever all the features are added to the
plot.