Fix - fix ValueErrors cause by the user switching models
Previsously value errors occured in the feature space callback as, when
a user switched models, if the dropdown had feature names selected from
a previous model then they would not exist in the new data frame.

A try except statement was used to catch this error and return an empty
scatter graph. The dropdown values are then updated to contain the
feature names in the new data frame, resolving the issue.