feat - Add unit testing framework
Unit tests can now be created in the appropriate directory in tests/ and
they can be run to asses line coverage.

A new singleton AppInstance was created to store the dash app. This was
required in order to change the way callbacks were added. Previously,
inner functions were used, which made the callbacks inaccessible for
testing. Now callbacks are imported directly into the app.py file, which
has made the SystemCallbacks.py file redundant.